refactor(jobs): extract refresh.go into modular components
Transform 608-line monolithic refresh.go into 5 focused modules following
the separation of concerns principle from your architecture proposal.

**Structure:**
- refresh.go: 359 lines (-41%) - Pure orchestration with tracing
- types.go: 106 lines - Interfaces (Logger, OpenWebIFClient, MetricsRecorder)
                        DTOs (Options, Deps, Artifacts, RefreshStats)
- validate.go: 80 lines - Config validation, filename sanitization, concurrency clamping
- write.go: 109 lines - Atomic file writes (M3U, XMLTV) with temp-file-rename pattern
- fetch.go: 163 lines - EPG collection with bounded concurrency, retry logic

**Benefits:**
✓ Each module has single responsibility
✓ Functions are now independently testable
✓ Atomic writes prevent partial file corruption
✓ Clear separation: orchestration vs I/O vs computation
✓ Easier to maintain and extend

**Testing:**
✓ All jobs tests pass (815/815 total)
✓ No behavioral changes
✓ Compilation verified

Next steps: Extract playlist building, add fakes for unit testing
