+++ internal/jobs/refresh.go
@@ -5,15 +5,11 @@ package jobs
 
 import (
 	"context"
-	"crypto/sha256"
-	"encoding/hex"
 	"errors"
 	"fmt"
-	"net/url"
 	"os"
 	"path/filepath"
 	"strings"
-	"sync"
 	"time"
 
 	xglog "github.com/ManuGH/xg2g/internal/log"
@@ -360,249 +356,4 @@ func Refresh(ctx context.Context, cfg Config) (*Status, error) {
 		Int("channels", status.Channels).
 		Msg("refresh completed")
 	return status, nil
-}
-
-// writeM3U safely writes the playlist to a temporary file and renames it on success.
-func writeM3U(ctx context.Context, path string, items []playlist.Item) error {
-	logger := xglog.FromContext(ctx)
-	dir := filepath.Dir(path)
-	tmpFile, err := os.CreateTemp(dir, "playlist-*.m3u.tmp")
-	if err != nil {
-		return fmt.Errorf("create temporary M3U file: %w", err)
-	}
-	// Defer a function to handle cleanup, logging any errors.
-	closed := false
-	defer func() {
-		if !closed {
-			if err := tmpFile.Close(); err != nil {
-				logger.Warn().Err(err).Str("path", tmpFile.Name()).Msg("failed to close temporary file on error path")
-			}
-		}
-		// Only remove the temp file if it still exists (i.e., rename failed).
-		if _, statErr := os.Stat(tmpFile.Name()); !os.IsNotExist(statErr) {
-			if err := os.Remove(tmpFile.Name()); err != nil {
-				logger.Warn().Err(err).Str("path", tmpFile.Name()).Msg("failed to remove temporary file")
-			}
-		}
-	}()
-
-	if err := playlist.WriteM3U(tmpFile, items); err != nil {
-		return fmt.Errorf("write to temporary M3U file: %w", err)
-	}
-
-	// Explicitly close before rename.
-	if err := tmpFile.Close(); err != nil {
-		return fmt.Errorf("close temporary M3U file before rename: %w", err)
-	}
-	closed = true
-
-	// Atomically rename the temporary file to the final destination
-	if err := os.Rename(tmpFile.Name(), path); err != nil {
-		return fmt.Errorf("rename temporary M3U file: %w", err)
-	}
-
-	return nil
-}
-
-// makeStableIDFromSRef creates a deterministic, collision-resistant tvg-id from a service reference.
-// Using a hash ensures the ID is stable even if the channel name changes and avoids issues
-// with special characters in the sRef.
-func makeStableIDFromSRef(sref string) string {
-	sum := sha256.Sum256([]byte(sref))
-	return "sref-" + hex.EncodeToString(sum[:])
-}
-
-func validateConfig(cfg Config) error {
-	// Use centralized validation package
-	v := validate.New()
-
-	v.URL("OWIBase", cfg.OWIBase, []string{"http", "https"})
-	v.Port("StreamPort", cfg.StreamPort)
-	v.Directory("DataDir", cfg.DataDir, false)
-
-	if !v.IsValid() {
-		return v.Err()
-	}
-
-	return nil
-}
-
-// collectEPGProgrammes fetches EPG data using per-service requests with bounded concurrency
-func collectEPGProgrammes(ctx context.Context, client *openwebif.Client, items []playlist.Item, cfg Config) []epg.Programme {
-	logger := xglog.FromContext(ctx)
-
-	// Clamp concurrency to sane bounds [1,10]
-	maxPar := cfg.EPGMaxConcurrency
-	if maxPar < 1 {
-		maxPar = 1
-	}
-	if maxPar > 10 {
-		maxPar = 10
-	}
-
-	// Worker pool semaphore
-	sem := make(chan struct{}, maxPar)
-	results := make(chan epgResult, len(items))
-	var wg sync.WaitGroup
-
-	// Schedule per-channel EPG fetches
-	for _, item := range items {
-		it := item // capture
-		// Validate sRef presence early to avoid spinning up a goroutine needlessly
-		if sref := extractSRefFromStreamURL(it.URL); sref == "" {
-			logger.Debug().Str("channel", it.Name).Msg("skipping EPG: could not extract sRef from stream URL")
-			continue
-		}
-
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			// Acquire semaphore
-			sem <- struct{}{}
-			defer func() { <-sem }()
-
-			// Deadline per request
-			reqCtx, cancel := context.WithTimeout(ctx, time.Duration(cfg.EPGTimeoutMS)*time.Millisecond)
-			defer cancel()
-
-			events, err := fetchEPGWithRetry(reqCtx, client, it.URL, cfg)
-			if err != nil {
-				logger.Debug().Err(err).
-					Str("channel", it.Name).
-					Str("tvg_id", it.TvgID).
-					Msg("EPG fetch failed for channel")
-				results <- epgResult{channelID: it.TvgID, events: nil, err: err}
-				return
-			}
-			results <- epgResult{channelID: it.TvgID, events: events, err: nil}
-		}()
-	}
-
-	// Close results when all goroutines complete
-	go func() {
-		wg.Wait()
-		close(results)
-	}()
-
-	// Aggregate results
-	var allProgrammes []epg.Programme
-	channelsWithData := 0
-
-	for res := range results {
-		if res.err != nil {
-			// already logged
-			continue
-		}
-		if len(res.events) > 0 {
-			channelsWithData++
-		}
-		progs := epg.ProgrammesFromEPG(res.events, res.channelID)
-		if len(progs) > 0 {
-			allProgrammes = append(allProgrammes, progs...)
-		}
-	}
-
-	logger.Info().
-		Int("total_programmes", len(allProgrammes)).
-		Int("channels_with_data", channelsWithData).
-		Int("concurrency", maxPar).
-		Msg("EPG collected via service endpoints")
-
-	metrics.RecordEPGChannelSuccess(channelsWithData)
-
-	return allProgrammes
-}
-
-type epgResult struct {
-	channelID string
-	events    []openwebif.EPGEvent
-	err       error
-}
-
-func fetchEPGWithRetry(ctx context.Context, client *openwebif.Client, streamURL string, cfg Config) ([]openwebif.EPGEvent, error) {
-	// Extract sRef from streamURL - adjust based on your stream URL format
-	sRef := extractSRefFromStreamURL(streamURL)
-	if sRef == "" {
-		return nil, fmt.Errorf("cannot extract sRef from stream URL: %s", streamURL)
-	}
-
-	var lastErr error
-	for attempt := 0; attempt <= cfg.EPGRetries; attempt++ {
-		if attempt > 0 {
-			// Exponential backoff
-			backoff := time.Duration(attempt*attempt*500) * time.Millisecond
-			select {
-			case <-time.After(backoff):
-			case <-ctx.Done():
-				return nil, ctx.Err()
-			}
-		}
-
-		events, err := client.GetEPG(ctx, sRef, cfg.EPGDays)
-		if err == nil {
-			return events, nil
-		}
-		lastErr = err
-	}
-
-	return nil, fmt.Errorf("EPG request failed after %d retries: %w", cfg.EPGRetries, lastErr)
-}
-
-// sanitizeFilename sanitizes a playlist filename to prevent path traversal attacks
-func sanitizeFilename(name string) (string, error) {
-	if name == "" {
-		return "playlist.m3u", nil
-	}
-
-	// Strip any directory components
-	base := filepath.Base(name)
-
-	// Reject if still contains traversal
-	if strings.Contains(base, "..") {
-		return "", fmt.Errorf("invalid filename: contains traversal")
-	}
-
-	// Clean the filename
-	cleaned := filepath.Clean(base)
-
-	// Ensure it's local
-	if !filepath.IsLocal(cleaned) {
-		return "", fmt.Errorf("invalid filename: not local")
-	}
-
-	// Validate extension
-	ext := filepath.Ext(cleaned)
-	if ext != ".m3u" && ext != ".m3u8" {
-		cleaned += ".m3u"
-	}
-
-	return cleaned, nil
-}
-
-// extractSRefFromStreamURL extracts service reference from stream URL
-func extractSRefFromStreamURL(streamURL string) string {
-	u, err := url.Parse(streamURL)
-	if err != nil {
-		return ""
-	}
-
-	// New format (direct service reference): http://host:port/1:0:19:132F:3EF:1:C00000:0:0:0:
-	// Service reference is in the path
-	path := strings.TrimPrefix(u.Path, "/")
-	if path != "" && strings.Contains(path, ":") {
-		return path
-	}
-
-	// Old format (fallback): http://host:port/web/stream.m3u?ref=ENCODED_SREF
-	encodedRef := u.Query().Get("ref")
-	if encodedRef == "" {
-		return ""
-	}
-
-	decodedRef, err := url.QueryUnescape(encodedRef)
-	if err != nil {
-		return ""
-	}
-
-	return decodedRef
 }
